Third attempt at Day 08 Puzzle 2
I had to verify that each path eventually looped back upon itself.
There are 6 starting points and each one is a full cycle that
immediately starts over itself.  So the paths are not too complex.  Now,
I need to extract common factors to get the correct convergence.

Ruby's `prime` library to the rescue!!

Luckily, all path lengths are prime multiples of 293.  The solution
becomes the product of all prime factors and 293.